uart: convert poll to libevent
Rather that use a thread per UART and use blocking I/O for each UART the
code has been switched to an event driven model as supplied by libevent.
This should supply faster responses with less CPU utilization which
means it will behave closer to that of real hardware since it will be
more similar to an interrupt style model.