main: add merchanism to control UARTs on/off and path
Added a mechanism to allow the user to disable or enable interfaces to
UARTs from the command line. The interface also allows users to override
the default path for the TTY used for the UART so that it may be
programmatically controlled. The interface was implemented hopefully
generically enought to allow future peripherals to be configured as
well (e.g. -p uart0=off -p uart1=/tmp/uart1). This change also changed
the option to specify the storage of the flash memory to -s from -p.